+++ scripts/init.ts
@@ -5,7 +5,7 @@ import {
   CREATE_ITEM_TABLE,
   CREATE_PLAYER_NAME_CHANGE_TABLE,
   CREATE_PLAYER_TABLE,
-} from "./db.mts";
+} from "./db";
 
 await sql.unsafe(CREATE_ITEM_TABLE);
 await sql.unsafe(CREATE_PLAYER_TABLE);
